Prevent unassigned objects
To improve compatibility and reduce errors, as well as follow best
practices, the classes that used their __construct() method to
perform all of their tasks when the object is created have been
refactored.

Now, when an object from one of those classes is instantiated, it is
assigned to a variable. Then, a method of that object is called to
run all the processes containd in the class.